requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic-settings==2.5.2
email-validator==2.2.0
pytest==8.3.3
httpx==0.27.2